Remove Results from `ast.rs`
Since my grammar is currently pretty messy, it is ambiguous whether a
parsing operation fails, ie expected something that wasn't there, or
just failed to find any matching patterns at all. Previously I tried to
guess at what was a fail and what was a miss, but this caused difficult
to catch errors. The current fix is just to remove the ambiguity,
returning only an Option, but the long term solution will be reworking
the grammar to be explicitly LL(1).